// nuestra_señora_del_rosario/src/types/GuardianType.ts
export interface Guardian {
    id_Guardian: number;      // ID del guardián
    name_GD: string;          // Nombre del guardián
    lastname1_GD: string;     // Primer apellido del guardián
    lastname2_GD: string;     // Segundo apellido del guardián
    cedula_GD: string;        // Cédula del guardián
    email_GD: string;         // Correo electrónico del guardián
    phone_GD: string;         // Teléfono del guardián
  }
  